Move tax breakdown objects
