+++ packages/core/src/Base/ValueObjects/TaxBreakdownAmount.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Lunar\Base\DataTransferObjects;
+namespace Lunar\Base\ValueObjects;
 
 use Lunar\DataTypes\Price;
