+++ packages/core/src/Drivers/SystemTaxDriver.php
@@ -4,8 +4,8 @@ namespace Lunar\Drivers;
 
 use Lunar\Actions\Taxes\GetTaxZone;
 use Lunar\Base\Addressable;
-use Lunar\Base\DataTransferObjects\TaxBreakdown;
-use Lunar\Base\DataTransferObjects\TaxBreakdownAmount;
+use Lunar\Base\ValueObjects\TaxBreakdown;
+use Lunar\Base\ValueObjects\TaxBreakdownAmount;
 use Lunar\Base\Purchasable;
 use Lunar\Base\TaxDriver;
 use Lunar\DataTypes\Price;
