+++ packages/core/src/Models/CartAddress.php
@@ -5,7 +5,7 @@ namespace Lunar\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Lunar\Base\Addressable;
 use Lunar\Base\BaseModel;
-use Lunar\Base\DataTransferObjects\TaxBreakdown;
+use Lunar\Base\ValueObjects\TaxBreakdown;
 use Lunar\Base\Traits\HasMacros;
 use Lunar\Base\Traits\LogsActivity;
 use Lunar\Database\Factories\CartAddressFactory;
@@ -49,7 +49,7 @@ class CartAddress extends BaseModel implements Addressable
     /**
      * The tax breakdown.
      *
-     * @var \Lunar\Base\DataTransferObjects\TaxBreakdown
+     * @var \Lunar\Base\ValueObjects\TaxBreakdown
      */
     public TaxBreakdown $taxBreakdown;
